minor code-style improvement
